LUCENE-3298: FSTs can now be larger than 2GB, have more than 2B nodes
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1433026 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/util/fst/ForwardBytesReader.java
@@ -46,13 +46,13 @@ final class ForwardBytesReader extends FST.BytesReader {
   }
 
   @Override
-  public int getPosition() {
+  public long getPosition() {
     return pos;
   }
 
   @Override
-  public void setPosition(int pos) {
-    this.pos = pos;
+  public void setPosition(long pos) {
+    this.pos = (int) pos;
   }
 
   @Override
