+++ lucene/core/src/java/org/apache/lucene/util/fst/ReverseBytesReader.java
@@ -44,13 +44,13 @@ final class ReverseBytesReader extends FST.BytesReader {
   }
 
   @Override
-  public int getPosition() {
+  public long getPosition() {
     return pos;
   }
 
   @Override
-  public void setPosition(int pos) {
-    this.pos = pos;
+  public void setPosition(long pos) {
+    this.pos = (int) pos;
   }
 
   @Override
